requirements.txt
Django==4.0
psycopg2-binary==2.9.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
django-filter==21.1